requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.6
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.37.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
plotly==5.23.0
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
replicate==0.25.2
requests==2.32.2
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.10.0
sortedcontainers==2.4.0
streamlit==1.34.0
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
torch==2.3.0
tornado==6.4.1
tqdm==4.66.3
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.0